Speed totem and Speed Leader options. Fixed build ordering bug. Maybe speed improvements, changed up locking, collecting and a couple of comps. Decimal sorting support. More CSV stats. Build timer. Streamlined sort points into one function. Added some helpers for Stats for leader skill functionality. Conditionally serializes members in Builds, smaller saves. Extra null checks on a bunch of stuff. Stores 'use equipped' in config.

+++ RuneOptim/Program.cs
@@ -442,7 +442,7 @@ namespace RuneOptim
             return runesSlot;
         }
 
-        public static Loadout MakeBuild(Monster mon, IEnumerable<Rune>[] runesSlot, Func<Stats, int> sort, Predicate<RuneSet[]> reqsets = null, Predicate<Stats> minimum = null)
+        public static Loadout MakeBuild(Monster mon, IEnumerable<Rune>[] runesSlot, Func<Stats, double> sort, Predicate<RuneSet[]> reqsets = null, Predicate<Stats> minimum = null)
         {
 
             System.Collections.Generic.SynchronizedCollection<Monster> tests = new SynchronizedCollection<Monster>();//new List<Monster>();
